refactor: move hcaptcha guard to hcaptcha module
